requirements.txt
pillow==11.3.0
pillow_heif==1.1.0
